refactor(reader): address review — harden selection tracker, broaden reflow trigger, dedup test harness
- Move the injected WebView scripts to ReaderWebViewScripts.kt (matches the
  TypographyOverride / FootnoteAnchorBridge convention).
- Selection tracker always overwrites window.__riffleSelSpan on a fresh
  non-collapsed selection (empty when no sentence-span ancestor), so selecting
  outside a narrated sentence falls back to the chapter position instead of
  reusing a previous selection's stale id.
- Re-apply decorations after ANY layout-affecting formatting change, not just
  orientation: rememberReflowReapplyGeneration now keys on formattingPrefs, so
  font-size/margin/line-spacing reflows also reposition a paused highlight.
- Extract the shared real-WebView test harness (withWebViewFixture/evalSync)
  used by both reader WebView-script tests; add a regression test for the
  stale-id reset.
